tests: replace old interrupt suite with typed inline-snapshot file (renamed to test_interrupts.py)
